requirements.txt
supabase
python-dotenv
openai


jinja2

requests